pyproject.toml
[tool.poetry]
name = "kmd"
# Keep this a dev version, as the dynamic versioning plugin is used for actual release versions:
version = "0.0.0.dev1"
description = "A command line for knowledge exploration."
authors = ["Joshua Levy <joshua@cal.berkeley.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
deepgram-sdk = "^3.5.0"
strif = "^1.0.1"
marko = "^2.1.2"
regex = "^2024.7.24"
openai = "^1.42.0"
cachetools = "^5.5.0"
requests = "^2.32.3"
pydub = "^0.25.1"
yt-dlp = "^2024.8.6"
assertpy = "^1.1"
ruamel-yaml = "^0.18.6"
python-slugify = "^8.0.4"
inflect = "^7.3.1"
justext = "^3.0.1"
lxml-html-clean = "^0.1.1"
cssselect = "^1.2.0"
weasyprint = "^62.3"
jinja2 = "^3.1.4"
textual = "^0.58.1"
textual-dev = "^1.5.1"
xonsh = "^0.17.0"
pygments = "^2.18.0"
humanize = "^4.10.0"
spacy = "^3.7.6"
litellm = "^1.44.1"
prompt-toolkit = "^3.0.47"
setproctitle = "^1.3.3"
tenacity = "^8.5.0"
nlpia2-wikipedia = "^1.5.17"
lazyasd = "^0.1.4"
chromadb = "^0.5.3"
llama-index-core = "^0.10.51"
llama-index-llms-openai = "^0.1.24"
llama-index-vector-stores-chroma = "^0.1.10"
llama-index-embeddings-openai = "^0.1.10"
python-dotenv = "^1.0.1"
curl-cffi = "^0.7.1"
thefuzz = "^0.22.1"
tiktoken = "^0.7.0"
scipy = "^1.14.1"
ripgrepy = "^2.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.2"
ruff = "^0.4.10"
usort = "^1.0.8.post1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Auto-generated version in build, based on tag or commit.
# https://sam.hooke.me/note/2023/08/poetry-automatically-generated-package-version-from-git-commit/
[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
pattern  = "^(?P<base>\\d+\\.\\d+\\.\\d+)(-?((?P<stage>[a-zA-Z]+)\\.?(?P<revision>\\d+)?))?"
format-jinja = """
    {%- if distance == 0 -%}
        {{- base -}}
    {%- else -%}
        {{- base }}.dev{{ distance }}+{{commit}}
    {%- endif -%}
"""

[tool.poetry.scripts]
kmd = "kmd.main:main"
lint = "lint:main"
test = "pytest:main"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
ignore = ["E402", "E731", "E712"]

[tool.pytest.ini_options]
python_files = ["*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = [
    "kmd",
    "tests",
]
norecursedirs = ["kmd/xontrib"]
filterwarnings = [
    "ignore::DeprecationWarning"
]
